Fix builds for Linux
