+++ component/libfuse/libfuse.go
@@ -36,8 +36,6 @@ import (
 	"github.com/Seagate/cloudfuse/common/log"
 	"github.com/Seagate/cloudfuse/internal"
 	"github.com/Seagate/cloudfuse/internal/stats_manager"
-
-	"github.com/winfsp/cgofuse/fuse"
 )
 
 /* NOTES:
@@ -46,33 +44,6 @@ import (
    - Order of calls : Constructor -> Configure -> Start ..... -> Stop
    - To read any new setting from config file follow the Configure method default comments
 */
-
-// Libfuse holds the settings and information for the FUSE component.
-type Libfuse struct {
-	internal.BaseComponent
-	host                  *fuse.FileSystemHost
-	mountPath             string
-	dirPermission         uint
-	filePermission        uint
-	readOnly              bool
-	attributeExpiration   uint32
-	entryExpiration       uint32
-	negativeTimeout       uint32
-	allowOther            bool
-	allowRoot             bool
-	ownerUID              uint32
-	ownerGID              uint32
-	traceEnable           bool
-	extensionPath         string
-	disableWritebackCache bool
-	ignoreOpenFlags       bool
-	nonEmptyMount         bool
-	networkShare          bool // Run as a network file share on Windows
-	lsFlags               common.BitMap16
-	maxFuseThreads        uint32
-	directIO              bool
-	umask                 uint32
-}
 
 // To support pagination in readdir calls this structure holds a block of items for a given directory
 type dirChildCache struct {
